backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.35
openai>=1.57.0,<2.0.0
httpx<0.28.0
pinecone==5.3.1
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
langchain-text-splitters==0.3.0
pdfplumber==0.11.4
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0
